+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## latest
 
+- Added support for workers in snapshot mode [#296](https://github.com/precice/micro-manager/pull/296)
 - Added check for number of workers and renamed method in `coupling.py` [#295](https://github.com/precice/micro-manager/pull/295)
 - Refactored interaction with preCICE and domain decomposition into `CouplingHandler`, `Profiler` and `DomainDecomposition` hierarchy [#289](https://github.com/precice/micro-manager/pull/289)
 - Refactored and generalized the methods in `p2p.py`. Added `MPIHandler` to manage all MPI related aspects. [#287](https://github.com/precice/micro-manager/pull/287)